+++ src/zope/app/apidoc/codemodule/tests.py
@@ -107,6 +107,20 @@ class TestModule(unittest.TestCase):
         self.assertEqual(mod['app']._module, zope.app)
         self.assertEqual(mod['app']['apidoc']._module, zope.app.apidoc)
 
+    def test_module_with_file_of_none(self):
+        # Sometimes namespace packages have this,
+        # especially on Python 3.7.
+        class Mod(object):
+            __file__ = None
+            __name__ = 'name'
+
+        mod = Mod()
+        mod.__doc__ = 'A module'
+
+        inst = Module(None, 'name', mod)
+        self.assertEqual(mod.__doc__, inst.getDocString())
+
+
 class TestZCML(unittest.TestCase):
 
     def setUp(self):