+++ CHANGES.rst
@@ -7,7 +7,8 @@
 
 - Host documentation at https://zopeappapidoc.readthedocs.io/
 
-- Add arguments to ``bin/static-apidoc`` for loading a specific ZCML file.
+- `#13 <https://github.com/zopefoundation/zope.app.apidoc/pull/13/>`_ Add argument to ``bin/static-apidoc`` for loading a specific ZCML file. To use this feature, the ZCML file you specify needs to 
+    <include package='zope.app.apidoc' file='ftesting.zcml' condition='have static-apidoc' />
 
 4.0.0 (2017-05-25)
 ==================