+++ CHANGES.rst
@@ -9,7 +9,7 @@
 
 - Add argument to ``static-apidoc`` for loading a specific ZCML file. To use this feature, the ZCML file you specify needs to 
   establish a working Zope 3 publication environment. The easiest way to do so is to include this line in the ZCML:
-  ``<include package='zope.app.apidoc' file='ftesting.zcml' condition='have static-apidoc' />``. 
+  ``<include package='zope.app.apidoc' file='static.zcml' condition='have static-apidoc' />``. 
   See `PR #13 <https://github.com/zopefoundation/zope.app.apidoc/pull/13/>`_.
 
 4.0.0 (2017-05-25)